requirements.txt
cffi==1.14.4
cryptography==3.3.1
Pillow==8.0.1
pycparser==2.20
six==1.15.0
